+++ tests/integration_test/conftest.py
@@ -2,102 +2,30 @@
 Fixtures for integration testing of the servers and client.
 """
 
-import pytest
+import json
 import os
 import random
-import subprocess
-import json
 import shutil
-import sys
 import socket
-
-from xprocess import ProcessStarter
-from socket import gethostname
+import subprocess
+import sys
 from pathlib import Path
+from socket import gethostname
+from subprocess import run
+
+import pytest
 from pydantic import BaseModel
+from xprocess import ProcessStarter
+
 from hera_librarian import LibrarianClient
+
+from ..server import Server, server_setup
 
 DATABASE_PATH = None
 SERVER_LOG_PATH = None
 
-SECONDARY_DATABASE_PATH = None
 
-class Server(BaseModel):
-    id: int
-    base_path: Path
-    staging_directory: Path
-    store_directory: Path
-    database: Path
-    LIBRARIAN_CONFIG_PATH: str
-    SQLALCHEMY_DATABASE_URI: str
-    PORT: str
-    ADD_STORES: str
-    process: str | None = None
-
-
-def server_setup(tmp_path_factory) -> Server:
-    """
-    Sets up a server.
-    """
-
-    librarian_config_path = str(Path("./tests/mock_config.json").resolve())
-
-    server_id_and_port = random.randint(1000, 20000)
-
-    # Check if the port is available. If not, increment until it is.
-    while (
-        socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(
-            (gethostname(), server_id_and_port)
-        )
-        == 0
-    ):
-        server_id_and_port += 1
-
-    tmp_path = tmp_path_factory.mktemp(f"server_{server_id_and_port}")
-
-    database = tmp_path / f"database_{server_id_and_port}.sqlite"
-
-    # Create the other server settings
-    staging_directory = tmp_path / f"staging_{server_id_and_port}"
-    staging_directory.mkdir()
-
-    store_directory = tmp_path / f"store_{server_id_and_port}"
-    store_directory.mkdir()
-
-    store_config = [
-        {
-            "store_name": "test_store",
-            "store_type": "local",
-            "ingestable": True,
-            "store_data": {
-                "staging_path": str(staging_directory),
-                "store_path": str(store_directory),
-            },
-            "transfer_manager_data": {
-                "local": {
-                    "available": "true",
-                    "hostname": gethostname(),
-                }
-            },
-        }
-    ]
-
-    add_stores = json.dumps(store_config)
-
-    return Server(
-        id=server_id_and_port,
-        base_path=tmp_path,
-        staging_directory=staging_directory,
-        store_directory=store_directory,
-        database=database,
-        LIBRARIAN_CONFIG_PATH=librarian_config_path,
-        SQLALCHEMY_DATABASE_URI=f"sqlite:///{database}",
-        PORT=str(server_id_and_port),
-        ADD_STORES=add_stores,
-    )
-
-
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="package")
 def server(xprocess, tmp_path_factory, request):
     """
     Starts a single server with pytest-xprocess.
@@ -133,8 +61,12 @@ def server(xprocess, tmp_path_factory, request):
 
 def pytest_terminal_summary(terminalreporter, exitstatus, config):
     terminalreporter.section("integration test temporary files")
-    terminalreporter.write_line("\033[1m" + "Server log: " + "\033[0m" + str(SERVER_LOG_PATH))
-    terminalreporter.write_line("\033[1m" + "Database: " + "\033[0m" + str(DATABASE_PATH))
+    terminalreporter.write_line(
+        "\033[1m" + "Server log: " + "\033[0m" + str(SERVER_LOG_PATH)
+    )
+    terminalreporter.write_line(
+        "\033[1m" + "Database: " + "\033[0m" + str(DATABASE_PATH)
+    )
 
 
 @pytest.fixture
@@ -154,22 +86,3 @@ def librarian_client(server) -> LibrarianClient:
     yield client
 
     del client
-
-
-@pytest.fixture
-def garbage_file(tmp_path) -> Path:
-    """
-    Returns a file filled with garbage at the path.
-    """
-
-    data = random.randbytes(1024)
-
-    path = tmp_path / "garbage_file.txt"
-
-    with open(path, "wb") as handle:
-        handle.write(data)
-
-    yield path
-
-    # Delete the file for good measure.
-    path.unlink()
