+++ tests/integration_test/test_configuration.py
@@ -3,6 +3,7 @@ Tests that we can successfully set up a server.
 """
 
 import requests
+
 from hera_librarian import LibrarianClient
 
 
@@ -11,7 +12,7 @@ def test_server(server):
     Tests that the server fixture is working.
     """
     assert server
-    
+
 
 def test_simple_ping(server):
     """
@@ -22,7 +23,7 @@ def test_simple_ping(server):
 
     # Just check we got something (even if its a 404)
     assert response.status_code
-    
+
 
 def test_ping_server(server):
     """
@@ -31,10 +32,7 @@ def test_ping_server(server):
 
     client = LibrarianClient(
         conn_name="test",
-        conn_config={
-            "url": f"http://localhost:{server.id}/",
-            "authenticator": None
-        }
+        conn_config={"url": f"http://localhost:{server.id}/", "authenticator": None},
     )
 
     assert client.ping()
