+++ tests/integration_test/test_upload_file.py
@@ -2,16 +2,13 @@
 Tests that you can successfully upload files to the server.
 """
 
-from pathlib import Path
 import sqlite3
+from pathlib import Path
+
 
 def test_upload_simple(librarian_client, garbage_file, server):
     # Perform the upload
-    librarian_client.upload_file(
-        garbage_file,
-        Path("test_file"),
-        null_obsid=True
-    )
+    librarian_client.upload_file(garbage_file, Path("test_file"), null_obsid=True)
 
     # Check we got it!
     # TODO: Implement that check within the librarian client (i.e. the ability to search db)
@@ -35,15 +32,15 @@ def test_upload_simple(librarian_client, garbage_file, server):
 
 def test_upload_file_to_unique_directory(librarian_client, garbage_file, server):
     librarian_client.upload_file(
-        garbage_file,
-        Path("test_directory/test_file"),
-        null_obsid=True
+        garbage_file, Path("test_directory/test_file"), null_obsid=True
     )
 
     # Check we got it (by manually verifying)
     conn = sqlite3.connect(server.database)
     c = conn.cursor()
-    res = c.execute("SELECT path FROM instances WHERE file_name='test_directory/test_file'")
+    res = c.execute(
+        "SELECT path FROM instances WHERE file_name='test_directory/test_file'"
+    )
     real_file_path = Path(res.fetchone()[0])
 
     assert real_file_path.exists()
@@ -55,5 +52,3 @@ def test_upload_file_to_unique_directory(librarian_client, garbage_file, server)
         garbage_file_contents = handle.read()
 
     assert real_file_contents == garbage_file_contents
-
-
